+++ projects/services-layers/src/lib/types/Layers.ts
@@ -165,6 +165,14 @@ export interface ILayerOptions {
    * If class 'hide' is included in the string, the layer is not shown in the UI - this can probably bring side effects when Layers are reordered, because the hidden layers could be moved on top off all!
    */
   cssClass?: string;
+  /**
+   * If the Layer consists of multiple layers but only one should be shown in the LayerControl
+   *
+   * If you use this then Layer.type must be 'custom' on this Layer. The merged Layers are overwritten for some attributes.
+   *
+   * Higher indexes get drawn above lower indexes
+   */
+  mergedLayers?: Layer[];
 }
 
 export interface ILayerDimensions extends IAnyObject {
@@ -276,6 +284,7 @@ export class Layer implements ILayerOptions {
   crossOrigin?: CrossOriginType;
   expanded = false;
   cssClass?: string;
+  mergedLayers?: Layer[];
 
   constructor(options: ILayerOptions) {
     Object.assign(this, options);